Use long cast while map.put()

+++ bvm/ballerina-runtime/src/main/java/org/ballerinalang/jvm/values/TableValueImpl.java
@@ -51,7 +51,6 @@ import static org.ballerinalang.jvm.util.BLangConstants.TABLE_LANG_LIB;
 import static org.ballerinalang.jvm.util.exceptions.BallerinaErrorReasons.OPERATION_NOT_SUPPORTED_IDENTIFIER;
 import static org.ballerinalang.jvm.util.exceptions.BallerinaErrorReasons.TABLE_HAS_A_VALUE_FOR_KEY_ERROR;
 import static org.ballerinalang.jvm.util.exceptions.BallerinaErrorReasons.TABLE_KEY_NOT_FOUND_ERROR;
-import static org.ballerinalang.jvm.util.exceptions.BallerinaErrorReasons.VALUE_INCONSISTENT_WITH_TABLE_TYPE_ERROR;
 
 /**
  * The runtime representation of table.
@@ -384,8 +383,8 @@ public class TableValueImpl<K, V> implements TableValue<K, V> {
         public V putData(V data) {
             Map.Entry<K, V> entry = new AbstractMap.SimpleEntry(data, data);
             UUID uuid = UUID.randomUUID();
-            entries.put(uuid.hashCode(), entry);
-            return values.put(uuid.hashCode(), data);
+            entries.put((long) uuid.hashCode(), entry);
+            return values.put((long) uuid.hashCode(), data);
         }
 
         public V remove(K key) {
